Power ups
Shield deactivates now after one shot against it. If ypu pick up a power
up object that is already active on you that time or number of lives
increases, power ups does not stack. The collision between your own
shield and yourself is NOT fixed! This has to be fixed in the collision
script where this collision is ignored.